requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.3
django-filter==25.1
django-jazzmin==3.0.1
django-mail-templated==2.6.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.3
drf-yasg==1.21.10
flake8==7.3.0
idna==3.10
importlib_resources==6.5.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
